Add command to allow all file premissions to local
Add a command to enable all file permissions to the local file and directory.

According to the directory synchronization between local and container,
the file owner and group in the container will be set to the UID (1000).

In the installation process of PrestaShop, it has a step to check file permission.

From the original official Docker for PrestaShop, it has defined file user
and group to be www-data.

So, these conditions result in PrestaShop can not be installed because it
has no permission to write file.

So, I decided to use a command `chmod` to enable all file permissions.
Although, enable all file permissions is not good for the security
aspect but in the development environment, it has not impact much.

It is another solution to solve this problem, using the UID from
the local to be the user to run docker but this solution requires to
modify every Dockerfile.